+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.0.32rc8] - UNRELEASED
+## [0.0.32rc10] - UNRELEASED
 
 ### Added